eth: define snap handler abstraction
